% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashmatrix.R
\name{dashmatrix}
\alias{dashmatrix}
\title{Dash Matrix}
\usage{
dashmatrix(data)
}
\arguments{
\item{data}{dataset with competitors as rows and judges as columns}
}
\value{
A matrix:
\item{dashmatrix}{matrix of number of placements}
}
\description{
Function to obtain the matrix of number of 1-1s, 1-2s, and so on.
}
\examples{
dashmatrix(testdata)
}
