% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankContestants.R
\name{rankContestants}
\alias{rankContestants}
\title{Ranking of Contestants}
\usage{
rankContestants(data)
}
\arguments{
\item{data}{dataset with competitors as rows and judges as columns}
}
\value{
A vector:
\item{finalranking}{final rankings of the competitors}
}
\description{
Function to rank contestants
}
\examples{
rankContestants(testdata)
}
