% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveTies.R
\name{resolveTies}
\alias{resolveTies}
\title{Resolve Ties}
\usage{
resolveTies(data, contestants, column)
}
\arguments{
\item{data}{dataset with competitors as rows and judges as columns}

\item{contestants}{vector with which contestant numbers to resolve ties for}

\item{column}{column  of the dash matrix to begin with}
}
\value{
A list:
\item{winnerfound}{method by which winner was found}
\item{winner}{vector with whom the winners were}
}
\description{
Function to resolve ties between competitors.
}
\examples{
resolveTies(testdata, c(1,2), 1)
}
