% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test Dataset}
\format{A data frame with 8 contestants (rows) and 5 judges (variables):
\describe{
  \item{J1}{rankings for Judge 1}
  \item{J2}{rankings for Judge 1}
  \item{J3}{rankings for Judge 1}
  \item{J4}{rankings for Judge 1}
  \item{J5}{rankings for Judge 1}
}}
\usage{
testdata
}
\description{
This synthetic dataset represents the placements of \emph{n} contestants (rows) by \emph{J} judges (columns).
}
\keyword{datasets}
